+++ src/MotionValidator.cpp
@@ -9,8 +9,8 @@ bool MotionValidator::checkMotion(const ob::State *s1, const ob::State *s2) cons
     const ob::RealVectorStateSpace::StateType *state1 = s1->as<ob::RealVectorStateSpace::StateType>();
     const ob::RealVectorStateSpace::StateType *state2 = s2->as<ob::RealVectorStateSpace::StateType>();
 
-    Eigen::Vector3f start(state1->values[0], state1->values[1], state1->values[2]);
-    Eigen::Vector3f end(state2->values[0], state2->values[1], state2->values[2]);
+    Eigen::Vector3d start(state1->values[0], state1->values[1], state1->values[2]);
+    Eigen::Vector3d end(state2->values[0], state2->values[1], state2->values[2]);
 
     return world->checkRayValid(start, end);
 }
@@ -20,21 +20,21 @@ bool MotionValidator::checkMotion(const ob::State *s1, const ob::State *s2, std:
     const ob::RealVectorStateSpace::StateType *state1 = s1->as<ob::RealVectorStateSpace::StateType>();
     const ob::RealVectorStateSpace::StateType *state2 = s2->as<ob::RealVectorStateSpace::StateType>();
 
-    Eigen::Vector3f start(state1->values[0], state1->values[1], state1->values[2]);
-    Eigen::Vector3f end(state2->values[0], state2->values[1], state2->values[2]);
+    Eigen::Vector3d start(state1->values[0], state1->values[1], state1->values[2]);
+    Eigen::Vector3d end(state2->values[0], state2->values[1], state2->values[2]);
 
     // run across ray in steps of delta
-    const float distance_delta = 0.1;
-    Eigen::Vector3f direction = (end - start).normalized();
-    Eigen::Vector3f current = start;
+    const double distance_delta = 0.1;
+    Eigen::Vector3d direction = (end - start).normalized();
+    Eigen::Vector3d current = start;
     const int steps = (end - start).norm() / distance_delta;
-    Eigen::Vector3f lastValidPoint = start;
+    Eigen::Vector3d lastValidPoint = start;
     for (int i = 0; i < steps; i++)
     {
         current = start + direction * i * distance_delta;
         if (!world->checkPointValidity(current))
         {
-           const float percentage = i / steps;
+           const double percentage = i / steps;
            lastValid.second = percentage;
 
            // converte lastValidPoint to state
