+++ src/OBB.cpp
@@ -5,20 +5,20 @@
 #include <string>
 #include <iostream>
 
-OBB::OBB(const Eigen::Vector3f &center, const Eigen::Vector3f &halfSize, const std::string &type) : center(center), halfSize(halfSize), type(type) {}
+OBB::OBB(const Eigen::Vector3d &center, const Eigen::Vector3d &halfSize, const std::string &type) : center(center), halfSize(halfSize), type(type) {}
 
-bool OBB::checkCollisionWithRay(const Eigen::Vector3f& start, const Eigen::Vector3f& end, const float inflateSize)const{
+bool OBB::checkCollisionWithRay(const Eigen::Vector3d& start, const Eigen::Vector3d& end, const double inflateSize)const{
     // Transform the ray into the OBB's local coordinate system
-    const Eigen::Vector3f localStart = rotation.transpose() * (start - center);
-    const Eigen::Vector3f localEnd = rotation.transpose() * (end - center);
-    const Eigen::Vector3f localDirection = (localEnd - localStart);
+    const Eigen::Vector3d localStart = rotation.transpose() * (start - center);
+    const Eigen::Vector3d localEnd = rotation.transpose() * (end - center);
+    const Eigen::Vector3d localDirection = (localEnd - localStart);
 
-    float tMin = 0;
-    float tMax = 1;
+    double tMin = 0;
+    double tMax = 1;
 
-    const Eigen::Vector3f inflatedHalfSize = halfSize + Eigen::Vector3f(inflateSize, inflateSize, inflateSize);
-    const Eigen::Vector3f box_min = -inflatedHalfSize;
-    const Eigen::Vector3f box_max = inflatedHalfSize;
+    const Eigen::Vector3d inflatedHalfSize = halfSize + Eigen::Vector3d(inflateSize, inflateSize, inflateSize);
+    const Eigen::Vector3d box_min = -inflatedHalfSize;
+    const Eigen::Vector3d box_max = inflatedHalfSize;
 
     for(int i = 0; i < 3; i++){
         if(abs(localDirection(i)) < 0.0001){
@@ -27,12 +27,12 @@ bool OBB::checkCollisionWithRay(const Eigen::Vector3f& start, const Eigen::Vecto
                 return false;
             }
         }else{
-            float invD = 1.0f / localDirection(i);
-            float t1 = (box_min(i) - localStart(i)) * invD;
-            float t2 = (box_max(i) - localStart(i)) * invD;
+            double invD = 1.0f / localDirection(i);
+            double t1 = (box_min(i) - localStart(i)) * invD;
+            double t2 = (box_max(i) - localStart(i)) * invD;
 
-            float tEntry = std::min(t1, t2);
-            float tExit = std::max(t1, t2);
+            double tEntry = std::min(t1, t2);
+            double tExit = std::max(t1, t2);
 
             tMin = std::max(tMin, tEntry);
             tMax = std::min(tMax, tExit);
@@ -46,10 +46,10 @@ bool OBB::checkCollisionWithRay(const Eigen::Vector3f& start, const Eigen::Vecto
     return 0 <= tMin && tMin <= 1 && 0 <= tMax && tMax <= 1;
 }
 
-bool OBB::checkCollisionWithPoint(const Eigen::Vector3f& point, const float inflateSize)const{
+bool OBB::checkCollisionWithPoint(const Eigen::Vector3d& point, const double inflateSize)const{
     // Transform the point into the OBB's local coordinate system
-    const Eigen::Vector3f localPoint = rotation.transpose() * (point - center);
-    const Eigen::Vector3f inflatedHalfSize = halfSize + Eigen::Vector3f(inflateSize, inflateSize, inflateSize);
+    const Eigen::Vector3d localPoint = rotation.transpose() * (point - center);
+    const Eigen::Vector3d inflatedHalfSize = halfSize + Eigen::Vector3d(inflateSize, inflateSize, inflateSize);
     
     // Check if the point is within the inflated bounding box
     return (std::abs(localPoint.x()) <= inflatedHalfSize.x()) &&
@@ -58,28 +58,28 @@ bool OBB::checkCollisionWithPoint(const Eigen::Vector3f& point, const float infl
 }
 
 
-box OBB::getAABB(const float inflateSize) const {
+box OBB::getAABB(const double inflateSize) const {
    // Get world fram AABB by rotating object into world frame and doing worst case estimation
    // Calculate the cornoes of the AABB based on the OBB center, half sizes and rotation
    // For this first create unit cube, scaled by halfSize
     
-    Eigen::Matrix3Xf corners(3, 8);
+    Eigen::Matrix3Xd corners(3, 8);
     corners << -1,  1,  1, -1, -1,  1,  1, -1,
                -1, -1,  1,  1, -1, -1,  1,  1,
                -1, -1, -1, -1,  1,  1,  1,  1;
     
-    Eigen::Matrix<float, 3, 8> halfSizesStacked = halfSize.replicate(1, 8);
+    Eigen::Matrix<double, 3, 8> halfSizesStacked = halfSize.replicate(1, 8);
     corners = corners.cwiseProduct(halfSizesStacked);
 
     // Rotate the corners
     // Todo, with or whithout transpose?
-    const Eigen::Matrix<float, 3, 8> centerStacked = center.replicate(1, 8);
-    const Eigen::Matrix3Xf cornersGlobalFrame = (rotation * corners)+ centerStacked;
-    Eigen::Vector3f minCorner = cornersGlobalFrame.rowwise().minCoeff();
-    Eigen::Vector3f maxCorner = cornersGlobalFrame.rowwise().maxCoeff();
+    const Eigen::Matrix<double, 3, 8> centerStacked = center.replicate(1, 8);
+    const Eigen::Matrix3Xd cornersGlobalFrame = (rotation * corners)+ centerStacked;
+    Eigen::Vector3d minCorner = cornersGlobalFrame.rowwise().minCoeff();
+    Eigen::Vector3d maxCorner = cornersGlobalFrame.rowwise().maxCoeff();
 
     // Inflate the AABB
-    const Eigen::Vector3f inflateSizeVec(inflateSize, inflateSize, inflateSize);
+    const Eigen::Vector3d inflateSizeVec(inflateSize, inflateSize, inflateSize);
     minCorner -= inflateSizeVec;
     maxCorner += inflateSizeVec;
